requirements.txt
altgraph==0.17.4
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
idna==3.10
packaging==24.2
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.2
python-dotenv==1.1.0
requests==2.32.3
setuptools==78.1.0
tabulate==0.9.0
urllib3==2.3.0
